+++ kopy-args/main/kotlin/org/gradle/kotlin/dsl/KopyCopyFunctions.kt
@@ -1,0 +1,5 @@
+package org.gradle.kotlin.dsl
+
+import com.javiersc.kotlin.kopy.args.KopyCopyFunctions
+
+public typealias KopyCopyFunctions = KopyCopyFunctions