+++ kopy-args/main/kotlin/org/gradle/kotlin/dsl/KopyTransformFunctions.kt
@@ -1,0 +1,5 @@
+package org.gradle.kotlin.dsl
+
+import com.javiersc.kotlin.kopy.args.KopyTransformFunctions
+
+public typealias KopyTransformFunction = KopyTransformFunctions