+++ kopy-args/main/kotlin/org/gradle/kotlin/dsl/KopyVisibility.kt
@@ -1,0 +1,3 @@
+package org.gradle.kotlin.dsl
+
+public typealias KopyVisibility = com.javiersc.kotlin.kopy.args.KopyVisibility