+++ kopy-runtime/common/main/kotlin/com/javiersc/kotlin/kopy/KopyOptIn.kt
@@ -4,7 +4,7 @@ package com.javiersc.kotlin.kopy
     message =
         """The Kopy plugin must be applied to use this API:
              • https://github.com/JavierSegoviaCordoba/kopy
-        """,
+        """
 )
 @Retention(AnnotationRetention.BINARY)
 @Target(AnnotationTarget.FUNCTION)